Updated UpdateLoadBalancer to create all new NBs, then rebuild them
NodeBalancerConfigCreateOptions includes a Nodes field to specify
nodes that should be used when creating the config, but that field is
actually ignored; the Linode API doesn't support it and won't create
nodes at config creation time. Until or unless that changes, the more
sane approach is to create a NodeBalancer config if one doesn't exist,
then always rebuild the "current" (new or existing) config.