allow inter-service communication from within cluster via external LB
This commit changes the LoadBalancer status we return from
EnsureLoadBalancer and GetLoadBalancer to only include the NodeBalancer
hostname and not the IPv4 address. Omitting the IP ensures
kube-proxy does not insert rules to hijack requests to the external LB
IP from within the cluster to resolve to the cluster IP for the
underlying pod. This proxying caused issues for users expecting the
NodeBalancer to include proxy-protocol headers. See #74.

Due to LoadBalancer statuses not including the LB IPv4 address, LB
discovery will now be done using the hostname.